requirements.txt
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
Django==5.1.1
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.2
ffmpeg==1.4
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
pillow==11.0.0
pycparser==2.22
PySocks==1.7.1
python-slugify==8.0.4
pytube==15.0.0
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
text-unidecode==1.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yt-dlp==2024.12.13
